backend/requirements.txt
asgiref==3.7.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-templated-mail==1.1.1
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
